ai-model/train/train_heart_model.py
# train_heart_model.py
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
import pickle

# Load dataset
df = pd.read_csv('../dataset/heart-disease.csv')

# Split features and target
X = df.drop('target', axis=1)
y = df['target']

# Train-test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Model training
model = RandomForestClassifier()
model.fit(X_train, y_train)

# Save the model
with open('../model/heart_model.pkl', 'wb') as f:
    pickle.dump(model, f)

print("✅ Heart disease model trained and saved.")
